requirements.txt
cycler==0.10.0
imageio==2.9.0
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.3.0.36
openslide-python==1.1.1
Pillow==8.0.1
pycocotools==2.0
scikit-learn==0.23.1
scipy==1.5.0
sklearn==0.0
statsmodels==0.11.1
torch==1.7.0
torchvision==0.8.1
tqdm==4.53.0
